+++ src/main/kotlin/cn/spacexc/bilibilisdk/sdk/article/info/remote/content/Tag.kt
@@ -1,9 +1,7 @@
 package cn.spacexc.bilibilisdk.sdk.article.info.remote.content
 
 
-import com.google.gson.annotations.SerializedName
-
 data class Tag(
-    val tid: Int,
+    val tid: Long,
     val name: String
 )
